// src/app/Models/iuser.ts
export interface IUser {
  firstName: string;
  secondName: string;
  lastName: string;
  birthDate: Date;
  mobileNo: number;
  email: string;
  whatsAppNo: number;
  password: string;
  // studentPhoneNumber:number ;
  // fatherPhoneNumber:number;
  // Religion: string;
  // birthday: Date;
  // Sex: string;
  // Governorate: string;
  // education: string;
  // address: string
  // password: number;
  // ConfirmPassword: number;
}